refactor(ddagen): Migrate project to typescript (#7)
