+++ .gitignore
@@ -23,8 +23,10 @@ yarn-debug.log*
 yarn-error.log*
 .pnpm-debug.log*
 
-# local env files
+# env files
 .env*.local
+.env
+next-env.d.ts
 
 # vercel
 .vercel
