+++ package.json
@@ -20,6 +20,7 @@
     "react-firebase-hooks": "^5.1.1"
   },
   "devDependencies": {
+    "@types/react": "18.0.28",
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.21",
     "tailwindcss": "^3.2.6"
