+++ src/locales/sv.ts
