--- src/pages/_app.js
@@ -1,11 +1,0 @@
-import Navbar from '@/components/Navbar';
-import '@/styles/globals.css'
-
-export default function App({ Component, pageProps }) {
-    return (
-        <>
-            <Navbar />
-            <Component {...pageProps} />
-        </>
-    );
-}
