+++ src/pages/_app.tsx
@@ -1,0 +1,12 @@
+import '@/styles/globals.css'
+import Navbar from '@/components/Navbar';
+import type { AppProps } from 'next/app'
+
+export default function App({ Component, pageProps }: AppProps) {
+    return (
+        <>
+            <Navbar />
+            <Component {...pageProps} />
+        </>
+    );
+}
