+++ src/pages/f%C3%B6retagsanm%C3%A4lan.tsx
@@ -1,0 +1,13 @@
+import CompanyForm from "@/components/CompanyForm";
+import { useLocale } from "@/locales";
+
+export default function CTA() {
+    const t = useLocale();
+
+    return (
+        <div>
+            <p>{t.nav.companyForm}</p>
+            <CompanyForm t={t} />
+        </div>
+    )
+}
