--- src/pages/f%C3%B6rf%C3%B6retag.js
@@ -1,14 +1,0 @@
-import en from "@/locales/en";
-import sv from "@/locales/sv";
-import { useRouter } from "next/router";
-
-export default function Companies() {
-    const router = useRouter();
-    const t = router.locale === 'sv' ? sv : en;
-
-    return (
-        <div>
-            <p>For Företag</p>
-        </div>
-    )
-}
