+++ src/pages/f%C3%B6rf%C3%B6retag.tsx
@@ -1,0 +1,11 @@
+import { useLocale } from "@/locales";
+
+export default function Companies() {
+    const t = useLocale()
+
+    return (
+        <div>
+            <p>{t.nav.forCompanies}</p>
+        </div>
+    )
+}
