+++ src/pages/f%C3%B6rstudenter.tsx
@@ -1,0 +1,11 @@
+import { useLocale } from "@/locales";
+
+export default function Students() {
+    const t = useLocale();
+
+    return (
+        <div>
+            <p>{t.nav.forStudents}</p>
+        </div>
+    )
+}
