+++ src/pages/faq.tsx
@@ -1,0 +1,11 @@
+import { useLocale } from "@/locales";
+
+export default function Faq() {
+    const t = useLocale();
+
+    return (
+        <div>
+            <p>FAQ</p>
+        </div>
+    )
+}
