+++ src/pages/katalog.tsx
@@ -1,0 +1,11 @@
+import { useLocale } from "@/locales";
+
+export default function Catalog() {
+    const t = useLocale();
+
+    return (
+        <div>
+            <p>Katalog</p>
+        </div>
+    )
+}
