+++ src/pages/kontakt.tsx
@@ -1,0 +1,11 @@
+import { useLocale } from "@/locales";
+
+export default function Contact() {
+    const t = useLocale();
+
+    return (
+        <div>
+            <p>Contact</p>
+        </div>
+    )
+}
