+++ src/pages/m%C3%A4ssan.tsx
@@ -1,0 +1,11 @@
+import { useLocale } from "@/locales";
+
+export default function About() {
+    const t = useLocale();
+
+    return (
+        <div>
+            <p>{t.about}</p>
+        </div>
+    )
+}
