+++ tailwind.config.js
@@ -1,8 +1,8 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    './src/pages/**/*.js',
-    './src/components/**/*.js'
+    './src/pages/**/*.tsx',
+    './src/components/**/*.tsx'
   ],
   theme: {
     extend: {
